refactor/#162: KidMySpace 및 WishArchive ViewModel 패키지 위치 수정 (viewmodel 패키지 추가)
